+++ projects/Mallard/src/navigation/helpers/base.tsx
@@ -1,13 +1,13 @@
 import React, { ReactElement } from 'react'
 import {
-    NavigationScreenProp,
-    NavigationInjectedProps,
     NavigationContainer,
+    NavigationInjectedProps,
+    NavigationScreenProp,
 } from 'react-navigation'
 import { routeNames } from 'src/navigation/routes'
-import { ArticleNavigator, PathToArticle } from 'src/screens/article-screen'
+import { PathToArticle, PathToIssue } from 'src/paths'
+import { ArticleNavigator } from 'src/screens/article-screen'
 import { Issue } from '../../../../common/src'
-import { PathToIssue } from 'src/paths'
 
 type NavigatorWrapper = ({ navigation }: NavigationInjectedProps) => JSX.Element
 export const addStaticRouter = (
